+++ src/Mammon/Utils/ObjectToInferredTypesConverter.cs
@@ -10,7 +10,7 @@ public class ObjectToInferredTypesConverter : JsonConverter<object>
             JsonTokenType.True => true,
             JsonTokenType.False => false,
             JsonTokenType.Number when reader.TryGetInt64(out long l) => l,
-            JsonTokenType.Number => reader.GetDouble(),
+            JsonTokenType.Number => reader.GetDecimal(),
             JsonTokenType.String when reader.TryGetDateTime(out DateTime datetime) => datetime,
             JsonTokenType.String => reader.GetString()!,
             _ => JsonDocument.ParseValue(ref reader).RootElement.Clone()
